* feature/new-product-entities: (28 commits)
  Changed ProductTest to use AssociateTypeCollectionFactory
  Added AssociateTypeCollectionFactory
  Changed name of RoleEntityCollectionFactory.createRole to RoleEntityCollectionFactory.createRoleCollection
  Added Associate to AssociateType relationship
  Changed Associate and AssociateType Entities to reflect database. Associate: Changed field description to bio AssociateType: Added description field
  Added Unit test for UserType Enum
  Added Associate table and join to Product table
  Created one to many relationship between Product and Image entities
  Created one to many relationship between Product and Image entities
  Image fix
  added Entityand Table image class added productId to image table
  Added AssociateType entity and test
  Added Image entity and test
  Changed ProductType id to ne type Long instead to int
  Removed redundant Public declaration in ProductDao and ProductService interfaces
  Changed id in Product entity to be of type Long instead of int
  Changed id in Genre to be of type Long instead of int
  Added Associate entity and unit test
  removed unused import from ProductTypeTest
  Added Genre entity
  ...

+++ src/main/java/com/waldronprojects/bookstore/controller/ProductController.java
@@ -41,7 +41,7 @@ public class ProductController {
 	}
 	
 	@GetMapping("/showFormForUpdate")
-	public String showFormForUpdate(@ModelAttribute("productId")int id,
+	public String showFormForUpdate(@ModelAttribute("productId")Long id,
 									Model model) {
 		Product product = productService.getProduct(id);
 		model.addAttribute("product", product);
@@ -49,7 +49,7 @@ public class ProductController {
 	}
 	
 	@GetMapping("/delete")
-	public String delete(@ModelAttribute("productId")int id) {
+	public String delete(@ModelAttribute("productId")Long id) {
 		productService.deleteProduct(id);
 		return "redirect:list";
 	}
